refactor(composition): group webui cluster under webui/ (dissection n10)
Move the WebUI facade and beta serve/middleware modules under a single
internal `webui/` module. The colliding `webui.rs` facade is renamed to
`webui/facade.rs`; the six `webui_*` middleware/serve files keep their
prefix. Crate-root `pub use` re-exports are repointed 1:1 so the public
API is unchanged (RebornWebuiBundle, build_webui_services now via
webui::facade; RateLimitConfigError via webui::webui_rate_limit; the
webui_serve item set via webui::webui_serve).

Pure move plus repaths — no logic changes.
