+++ crates/ironclaw_reborn_composition/src/webui/webui_operator_auth.rs
@@ -10,7 +10,7 @@ use axum::extract::Request;
 use axum::http::Method;
 use ironclaw_host_api::ingress::IngressRouteDescriptor;
 
-use crate::webui_route_match::{network_method_to_axum, parse_pattern, segments_match};
+use crate::webui::webui_route_match::{network_method_to_axum, parse_pattern, segments_match};
 
 #[derive(Debug, Clone)]
 struct OperatorWebuiConfigRoute {
