+++ crates/ironclaw_reborn_composition/src/webui/webui_ws_origin.rs
@@ -1,7 +1,7 @@
 //! Descriptor-driven `Origin` enforcement for WebChat v2 routes that
 //! declare a [`WebSocketOriginPolicy`] other than `NotApplicable`.
 //!
-//! The CORS layer composed by [`crate::webui_serve::webui_v2_app`]
+//! The CORS layer composed by [`crate::webui::webui_serve::webui_v2_app`]
 //! handles ordinary XHR pre-flight, but the browser does NOT issue a
 //! pre-flight before a WebSocket upgrade — it just opens a new
 //! connection and sends `Origin` directly on the upgrade request.
@@ -20,7 +20,7 @@ use axum::middleware::Next;
 use axum::response::{IntoResponse, Response};
 use ironclaw_host_api::ingress::{IngressRouteDescriptor, StreamingMode, WebSocketOriginPolicy};
 
-use crate::webui_route_match::{network_method_to_axum, parse_pattern, segments_match};
+use crate::webui::webui_route_match::{network_method_to_axum, parse_pattern, segments_match};
 
 #[derive(Debug, Clone)]
 struct WsRouteOriginRule {
